Inline Extensions.InvokeAsync methods into RemoteEndPoint
